Backend logic and frontend are refactored
