+++ User-service/src/main/java/Main.java
@@ -2,6 +2,6 @@ import frontend.Cli;
 
 public class Main {
     public static void main(String[] args) {
-        new Cli().appCycle();
+        new Cli().launch();
     }
 }
