+++ User-service/src/main/java/backend/HibernateUtil.java
@@ -7,7 +7,6 @@ import org.hibernate.boot.MetadataSources;
 import org.hibernate.boot.registry.StandardServiceRegistry;
 import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
 
-
 public class HibernateUtil {
 
     @Getter
