+++ User-service/src/main/java/backend/UserDao.java
@@ -1,28 +1,11 @@
 package backend;
 
-import frontend.ModeConstants;
 import org.hibernate.Session;
 import org.hibernate.query.Query;
 
 import java.util.List;
 
 public class UserDao {
-
-    public int saveUser(User user) {
-        org.hibernate.Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.getTransaction();
-            transaction.begin();
-            session.persist(user);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null && (transaction.isActive() || transaction.getRollbackOnly())) {
-                transaction.rollback();
-                return ModeConstants.ERROR;
-            }
-        }
-        return ModeConstants.OK;
-    }
 
     public User getUserById(String uuid) {
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
@@ -37,28 +20,29 @@ public class UserDao {
         }
     }
 
+    public int saveUser(User user) {
+        return handleTransaction(user, ModeConstants.SAVE);
+    }
+
     public int updateUser(User user) {
-        org.hibernate.Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.getTransaction();
-            transaction.begin();
-            session.merge(user);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null && (transaction.isActive() || transaction.getRollbackOnly())) {
-                transaction.rollback();
-                return ModeConstants.ERROR;
-            }
-        }
-        return ModeConstants.OK;
+        return handleTransaction(user, ModeConstants.UPDATE);
     }
 
     public int deleteUser(User user) {
+        return handleTransaction(user, ModeConstants.DELETE);
+    }
+
+    private int handleTransaction(User user, int mode) {
         org.hibernate.Transaction transaction = null;
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
             transaction = session.getTransaction();
             transaction.begin();
-            session.remove(user);
+            switch (mode) {
+                case ModeConstants.SAVE -> session.persist(user);
+                case ModeConstants.UPDATE -> session.merge(user);
+                case ModeConstants.DELETE -> session.remove(user);
+                default -> {}
+            }
             transaction.commit();
         } catch (Exception e) {
             if (transaction != null && (transaction.isActive() || transaction.getRollbackOnly())) {
